LLM Vendors rename
