+++ src/modules/llms/llm.types.ts
@@ -1,8 +1,8 @@
 import type React from 'react';
 
-import type { LLMOptionsOpenAI, SourceSetupOpenAI } from './openai/vendor';
+import type { LLMOptionsOpenAI, SourceSetupOpenAI } from './openai/openai.vendor';
 import type { OpenAI } from './openai/openai.types';
-import type { SourceSetupLocalAI } from './localai/vendor';
+import type { SourceSetupLocalAI } from './localai/localai.vendor';
 
 
 export type DLLMId = string;
